refactor code
- split task module into task, trial, vis modules
- export user classes and functions in init
- rename settings -> trial
