+++ .pre-commit-config.yaml
@@ -39,3 +39,8 @@ repos:
     hooks:
       - id: mypy
         args: [--ignore-missing-imports, --disallow-untyped-defs]
+
+  - repo: https://github.com/asottile/reorder_python_imports
+    rev: v3.8.2
+    hooks:
+      - id: reorder-python-imports
