+++ src/motor_task_prototype/__init__.py
@@ -1,1 +1,5 @@
+from motor_task_prototype.task import MotorTask
+from motor_task_prototype.trial import get_trial_from_user
+from motor_task_prototype.vis import display_results
+
 __version__ = "0.0.5"
